+++ src/main/java/lookIT/lookITspring/config/SpringConfig.java
@@ -1,5 +1,6 @@
 package lookIT.lookITspring.config;
 
+import java.util.Properties;
 import lookIT.lookITspring.repository.CollectionsRepository;
 import lookIT.lookITspring.repository.FriendTagsRepository;
 import lookIT.lookITspring.repository.FriendsRepository;
@@ -29,6 +30,8 @@ import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactor
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.JavaMailSenderImpl;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -38,9 +41,26 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 @EnableRedisRepositories
 public class SpringConfig {
     @Value("${spring.redis.host}")
-    private String host;
+    private String redis_host;
     @Value("${spring.redis.port}")
+    private int redis_port;
+
+    @Value("${mail.smtp.port}")
     private int port;
+    @Value("${mail.smtp.socketFactory.port}")
+    private int socketPort;
+    @Value("${mail.smtp.auth}")
+    private boolean auth;
+    @Value("${mail.smtp.starttls.enable}")
+    private boolean starttls;
+    @Value("${mail.smtp.starttls.required}")
+    private boolean startlls_required;
+    @Value("${mail.smtp.socketFactory.fallback}")
+    private boolean fallback;
+    @Value("${AdminMail.id}")
+    private String id;
+    @Value("${AdminMail.password}")
+    private String password;
 
     private final UserRepository userRepository;
     private final MemorySpotRepository memorySpotRepository;
@@ -77,7 +97,7 @@ public class SpringConfig {
     // lettuce
     @Bean
     public RedisConnectionFactory redisConnectionFactory() {
-        return new LettuceConnectionFactory(host, port);
+        return new LettuceConnectionFactory(redis_host, redis_port);
     }
 
     @Bean
@@ -100,8 +120,36 @@ public class SpringConfig {
     }
 
     @Bean
+    public JavaMailSender javaMailService() {
+        JavaMailSenderImpl javaMailSender = new JavaMailSenderImpl();
+        javaMailSender.setHost("smtp.gmail.com");
+        javaMailSender.setUsername(id);
+        javaMailSender.setPassword(password);
+        javaMailSender.setPort(port);
+        javaMailSender.setJavaMailProperties(getMailProperties());
+        javaMailSender.setDefaultEncoding("UTF-8");
+        return javaMailSender;
+    }
+
+    private Properties getMailProperties() {
+        Properties pt = new Properties();
+        pt.put("mail.smtp.socketFactory.port", socketPort);
+        pt.put("mail.smtp.auth", auth);
+        pt.put("mail.smtp.starttls.enable", starttls);
+        pt.put("mail.smtp.starttls.required", startlls_required);
+        pt.put("mail.smtp.socketFactory.fallback", fallback);
+        pt.put("mail.smtp.socketFactory.class", "javax.net.ssl.SSLSocketFactory");
+        return pt;
+    }
+
+    @Bean
+    public EmailService emailService() {
+        return new EmailService(javaMailService());
+    }
+
+    @Bean
     public UserService userService() {
-        return new UserService(userRepository, passwordEncoder(), jwtProvider(), redisTemplate(), new EmailService());
+        return new UserService(userRepository, passwordEncoder(), jwtProvider(), redisTemplate(), emailService());
     }
 
     @Bean
