+++ src/main/java/lookIT/lookITspring/service/EmailService.java
@@ -6,16 +6,16 @@ import javax.mail.Message.RecipientType;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.MailException;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 
-@Service
+@RequiredArgsConstructor
 public class EmailService {
 
-    @Autowired
-    JavaMailSender emailSender;
+    private final JavaMailSender emailSender;
 
     public static final String ePw = createKey();
